+++ src/baseobjects/objects/__init__.py
@@ -14,4 +14,3 @@ __email__ = __email__
 # Imports
 # Local Packages #
 from .automaticproperties import AutomaticProperties
-from .wrappers import StaticWrapper, DynamicWrapper
